+++ example/lib/main.dart
@@ -10,15 +10,15 @@
  *******************************************************************************
 */
 
-import 'package:flutter/material.dart';
-import 'package:intl/date_symbol_data_local.dart';
-import 'package:intl/find_locale.dart';
-import 'package:intl/intl.dart';
 import 'dart:async';
-
+import 'package:flutter/material.dart';
 import 'package:region_settings/region_settings.dart';
 
-void main() {
+late RegionSettings regionSettings;
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  regionSettings = await RegionSettings.getSettings();
   runApp(const MyApp());
 }
 
@@ -30,6 +30,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
+  String? _locale;
   TemperatureUnit? _temperatureUnits;
   bool? _usesMetricSystem;
   int? _firstDayOfWeek;
@@ -47,7 +48,8 @@ class _MyAppState extends State<MyApp> {
 
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> loadRegionSettings() async {
-    final RegionSettings regionSettings = await RegionSettings.getSettings();
+    regionSettings = await RegionSettings.getSettings();
+    String locale = regionSettings.locale;
     TemperatureUnit temperatureUnits = regionSettings.temperatureUnits;
     bool usesMetricSystem = regionSettings.usesMetricSystem;
     int firstDayOfWeek = regionSettings.firstDayOfWeek;
@@ -57,16 +59,13 @@ class _MyAppState extends State<MyApp> {
     String numberFormatInteger = regionSettings.numberFormat.integer;
     String numberFormatDecimal = regionSettings.numberFormat.decimal;
 
-    // Get default locale for DateFormat
-    await initializeDateFormatting();
-    Intl.defaultLocale = await findSystemLocale();
-
     // If the widget was removed from the tree while the asynchronous platform
     // message was in flight, we want to discard the reply rather than calling
     // setState to update our non-existent appearance.
     if (!mounted) return;
 
     setState(() {
+      _locale = locale;
       _temperatureUnits = temperatureUnits;
       _usesMetricSystem = usesMetricSystem;
       _firstDayOfWeek = firstDayOfWeek;
@@ -96,6 +95,11 @@ class _MyAppState extends State<MyApp> {
                   childAspectRatio: 4.0,
                   shrinkWrap: true,
                   children: [
+                    const Text('locale:'),
+                    Text(
+                      '$_locale',
+                      style: const TextStyle(fontWeight: FontWeight.bold),
+                    ),
                     const Text('temperatureUnits:'),
                     Text(
                       '$_temperatureUnits',
@@ -121,9 +125,12 @@ class _MyAppState extends State<MyApp> {
                       '$_dateFormatShort',
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
-                    const Text('Today as dateFormat.short:'),
+                    const Text('formatDate() as short:'),
                     Text(
-                      DateFormat(_dateFormatShort).format(DateTime.now()),
+                      regionSettings.formatDate(
+                        DateTime.now(),
+                        dateStyle: DateStyle.short,
+                      ),
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
                     const Text('dateFormat.medium:'),
@@ -131,9 +138,12 @@ class _MyAppState extends State<MyApp> {
                       '$_dateFormatMedium',
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
-                    const Text('Today as dateFormat.medium:'),
+                    const Text('formatDate() as medium:'),
                     Text(
-                      DateFormat(_dateFormatMedium).format(DateTime.now()),
+                      regionSettings.formatDate(
+                        DateTime.now(),
+                        dateStyle: DateStyle.medium,
+                      ),
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
                     const Text('dateFormat.long:'),
@@ -141,9 +151,12 @@ class _MyAppState extends State<MyApp> {
                       '$_dateFormatLong',
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
-                    const Text('Today as dateFormat.long:'),
+                    const Text('formatDate() as long:'),
                     Text(
-                      DateFormat(_dateFormatLong).format(DateTime.now()),
+                      regionSettings.formatDate(
+                        DateTime.now(),
+                        dateStyle: DateStyle.long,
+                      ),
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
                     const Text('numberFormat.integer:'),
@@ -151,11 +164,10 @@ class _MyAppState extends State<MyApp> {
                       '$_numberFormatInteger',
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
-                    const Text('Sample integer:'),
+                    const Text('formatNumber() as integer:'),
                     Text(
-                      _formatNumberWithPattern(
+                      regionSettings.formatNumber(
                         1234567,
-                        _numberFormatInteger,
                       ),
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
@@ -164,12 +176,12 @@ class _MyAppState extends State<MyApp> {
                       '$_numberFormatDecimal',
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
-                    const Text('Sample decimal:'),
+                    const Text('formatNumber() as decimal:'),
                     Text(
-                      _formatNumberWithPattern(
+                      regionSettings.formatNumber(
                         1234567.89,
-                        _numberFormatDecimal,
-                        asDecimal: true,
+                        decimalPlaces: 2,
+                        useGrouping: true,
                       ),
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
@@ -185,38 +197,5 @@ class _MyAppState extends State<MyApp> {
         ),
       ),
     );
-  }
-
-  // Convert number to string using a format pattern
-  String _formatNumberWithPattern(double number, String? pattern,
-      {bool asDecimal = false}) {
-    // Handle blank patterns
-    if (pattern == null) {
-      return number.toString();
-    }
-
-    // Convert the number to a string with expected precision and no decimal point
-    String numberStr =
-        number.toStringAsFixed(asDecimal ? 2 : 0).replaceAll('.', '');
-
-    // Throw error if pattern is too short
-    assert(numberStr.length <= '#'.allMatches(pattern).length);
-
-    // Iterate over the pattern and build formatted string
-    String result = '';
-    int numberIndex = numberStr.length - 1;
-    for (int patternIndex = pattern.length - 1;
-        patternIndex >= 0 && numberIndex >= 0;
-        patternIndex--) {
-      if (pattern[patternIndex] == '#') {
-        // Replace placeholder with digit
-        result = '${numberStr[numberIndex]}$result';
-        numberIndex--;
-      } else {
-        // Append non-numeric character unchanged
-        result = '${pattern[patternIndex]}$result';
-      }
-    }
-    return result;
   }
 }