+++ lib/region_settings.dart
@@ -12,6 +12,12 @@
 
 import 'region_settings_platform_interface.dart';
 
+import 'package:intl/date_symbol_data_local.dart';
+import 'package:intl/find_locale.dart';
+import 'package:intl/intl.dart';
+import 'package:intl/number_symbols_data.dart';
+import 'package:intl/number_symbols.dart';
+
 /// Temperature unit options enum.
 ///
 /// | enum | value |
@@ -37,6 +43,24 @@ enum TemperatureUnit {
   }
 }
 
+/// Date format styles enum.
+///
+/// This enum defines three styles of date formats that can be used with
+/// [RegionSettings.formatDate] or [RegionDateFormats]:
+/// * short
+/// * medium
+/// * long
+enum DateStyle {
+  /// Short date format, e.g. M/d/yy.
+  short,
+
+  /// Medium date format, e.g. MMM d, y.
+  medium,
+
+  /// Long date format, e.g. MMMM d, y.
+  long
+}
+
 /// Date format options.
 ///
 /// This class contains three string fields corresponding to three different
@@ -60,6 +84,18 @@ class RegionDateFormats {
 
   /// Long date format, e.g. MMMM d, y.
   final String long;
+
+  /// Get the date format pattern for a given DateStyle.
+  String pattern(DateStyle style) {
+    switch (style) {
+      case DateStyle.short:
+        return short;
+      case DateStyle.medium:
+        return medium;
+      case DateStyle.long:
+        return long;
+    }
+  }
 
   /// Output date format options as a formatted string.
   @override
@@ -108,12 +144,19 @@ class RegionNumberFormats {
 class RegionSettings {
   /// Constructs an instance of regional settings data.
   RegionSettings({
+    required this.locale,
     required this.temperatureUnits,
     required this.usesMetricSystem,
     required this.firstDayOfWeek,
     required this.dateFormat,
     required this.numberFormat,
+    required this.icuNumberFormat,
+    required this.decimalSeparator,
+    required this.groupSeparator,
   });
+
+  /// The locale for the region settings.
+  final String locale;
 
   /// Temperature unit options enum.
   final TemperatureUnit temperatureUnits;
@@ -130,19 +173,52 @@ class RegionSettings {
   /// Number format options.
   final RegionNumberFormats numberFormat;
 
+  /// ICU standard number format pattern.
+  ///
+  /// This is a string that can be used with
+  /// [intl NumberFormat](https://pub.dev/packages/intl) to format numbers.
+  final String icuNumberFormat;
+
+  /// The character used to separate decimal places in numbers.
+  final String decimalSeparator;
+
+  /// The character used to separate groups of thousands/hundreds in numbers.
+  final String groupSeparator;
+
+  /// Constants related to default format patterns.
+  static const String _defaultDecimalSeparator = '.';
+  static const String _defaultGroupSeparator = ',';
+  static const String _nonDigitRegex = r'[^#0]';
+  static const String _numberSymbolsName = 'regionsettings';
+
   /// Load all available regional settings from the device settings.
   ///
   /// Returns a [RegionSettings] object containing the current regional
   /// settings for temperature units, metric system use, first day of the week,
   /// date format patterns, and number format patterns.
   static Future<RegionSettings> getSettings() async {
+    // Get regional settings from platform
     TemperatureUnit temperatureUnits = await getTemperatureUnits();
     bool usesMetricSystem = await getUsesMetricSystem();
     int firstDayOfWeek = await getFirstDayOfWeek();
     List<String> dateFormatsList = await getDateFormatsList();
     List<String> numberFormatsList = await getNumberFormatsList();
+    String decimalSeparator =
+        _findLastNonDigit(numberFormatsList[1]) ?? _defaultDecimalSeparator;
+    String groupSeparator =
+        _findFirstNonDigit(numberFormatsList[1]) ?? _defaultGroupSeparator;
+    String icuNumberFormat = _getIcuNumberFormat(
+        numberFormatsList[1], decimalSeparator, groupSeparator);
+
+    // Initialize intl for date and number formatting
+    await initializeDateFormatting();
+    String locale = await findSystemLocale();
+    Intl.defaultLocale = locale;
+    _setNumberSymbols(
+        locale, icuNumberFormat, decimalSeparator, groupSeparator);
 
     return RegionSettings(
+      locale: locale,
       temperatureUnits: temperatureUnits,
       usesMetricSystem: usesMetricSystem,
       firstDayOfWeek: firstDayOfWeek,
@@ -155,7 +231,78 @@ class RegionSettings {
         integer: numberFormatsList[0],
         decimal: numberFormatsList[1],
       ),
+      icuNumberFormat: icuNumberFormat,
+      decimalSeparator: decimalSeparator,
+      groupSeparator: groupSeparator,
     );
+  }
+
+  /// Format a date using the regional settings.
+  ///
+  /// Returns the formatted [String] representation of a [DateTime] with the
+  /// plaform's locale and regional settings preferences applied. This is a
+  /// convenience method that uses the [intl DateFormat](https://pub.dev/packages/intl)
+  /// class to format the date without needing to manually specify the pattern.
+  ///
+  /// Before using this method, ensure that [getSettings] has been called to
+  /// load the plaform settings.
+  ///
+  /// Parameters:
+  /// * [date] - The date to format.
+  /// * [dateStyle] - Specify the date style. Defaults to [DateStyle.medium].
+  String formatDate(
+    DateTime date, {
+    DateStyle dateStyle = DateStyle.medium,
+  }) {
+    String pattern;
+    switch (dateStyle) {
+      case DateStyle.short:
+        pattern = dateFormat.short;
+        break;
+      case DateStyle.medium:
+        pattern = dateFormat.medium;
+        break;
+      case DateStyle.long:
+        pattern = dateFormat.long;
+        break;
+    }
+
+    DateFormat formatter = DateFormat(pattern, locale);
+    return formatter.format(date);
+  }
+
+  /// Format a number using the regional settings.
+  ///
+  /// Returns a formatted [String] representation of the number with the
+  /// plaform's regional settings preferences applied. Separator characters and
+  /// grouping styles are determined by the device's locale and region settings.
+  ///
+  /// Before using this method, ensure that [getSettings] has been called to
+  /// load the plaform settings.
+  ///
+  /// Parameters:
+  /// * [number] - The number to format.
+  /// * [decimalPlaces] - Specify decimal places to show. Defaults to 0.
+  /// * [useGrouping] - Separate into groups (e.g. thousands). Defaults to true.
+  String formatNumber(
+    double number, {
+    int decimalPlaces = 0,
+    bool useGrouping = true,
+  }) {
+    assert(decimalPlaces >= 0, 'decimalPlaces must be a non-negative integer');
+
+    // Ensure the number symbols are configured before we attempt to use them
+    _setNumberSymbols(
+        locale, icuNumberFormat, decimalSeparator, groupSeparator);
+
+    NumberFormat formatter = NumberFormat(icuNumberFormat, _numberSymbolsName);
+    formatter.minimumFractionDigits = decimalPlaces;
+    formatter.maximumFractionDigits = decimalPlaces;
+    if (!useGrouping) {
+      formatter.turnOffGrouping();
+    }
+
+    return formatter.format(number);
   }
 
   /// Get the temperature units from device settings.
@@ -204,9 +351,11 @@ class RegionSettings {
   ///
   /// ## Android Implementation
   ///
-  /// Android does not have an OS-level measurement system setting. Instead,
-  /// the plugin must guess the measurement system based on the device's
-  /// locale. The plugin considers the following countries to be non-Metric:
+  /// Android 16 and later include a separate Measurement System preference in
+  /// Regional preferences. If available, the plugin will use this setting. If
+  /// not available, the plugin must guess the measurement system based on the
+  /// device's locale. The plugin considers the following countries to be
+  /// non-Metric:
   /// * American Samoa (US)
   /// * Bahamas
   /// * Belize
@@ -406,15 +555,110 @@ class RegionSettings {
     return Future.value(numberFormatsList);
   }
 
+  /// Search a string for the first instance of any non-digit character.
+  ///
+  /// This method is private and used internally to parse platform patterns.
+  static String? _findFirstNonDigit(String text) {
+    RegExp nonDigit = RegExp(_nonDigitRegex);
+    Match? match = nonDigit.firstMatch(text);
+    return match?.group(0);
+  }
+
+  /// Search a string for the last instance of any non-digit character.
+  ///
+  /// This method is private and used internally to parse platform patterns.
+  static String? _findLastNonDigit(String text) {
+    RegExp nonDigit = RegExp(_nonDigitRegex);
+    int lastIndex = text.lastIndexOf(nonDigit);
+
+    if (lastIndex != -1) {
+      return text[lastIndex];
+    } else {
+      return null;
+    }
+  }
+
+  /// Determine the ICU standard number format pattern from a region pattern.
+  /// This is needed to convert platform number patterns into a format that can
+  /// be used with [intl NumberFormat](https://pub.dev/packages/intl), which
+  /// expects ICU style patterns where decimal and grouping separators are
+  /// standardized.
+  ///
+  /// This method is private and used internally to parse platform patterns.
+  static String _getIcuNumberFormat(
+    String formatPattern,
+    String decimalSeparator,
+    String groupSeparator,
+  ) {
+    // Use a map to store the replacement for each separator
+    final Map<String, String> replacements = {
+      groupSeparator: _defaultGroupSeparator,
+      decimalSeparator: _defaultDecimalSeparator,
+    };
+
+    // Create a regular expression that matches either separator character
+    final RegExp separatorRegExp = RegExp(
+      '(${RegExp.escape(groupSeparator)}|${RegExp.escape(decimalSeparator)})',
+    );
+
+    return formatPattern.replaceAllMapped(separatorRegExp, (match) {
+      String matchedSeparator = match.group(0)!;
+      // Return the corresponding replacement from the map
+      return replacements[matchedSeparator]!;
+    });
+  }
+
+  /// Create number formatting symbols using regional settings.
+  ///
+  /// This method sets up the number formatting symbols for the current locale
+  /// and overrides the decimal and group separators with the regional settings.
+  /// It is called during initialization to ensure that the number formatting
+  /// symbols are correctly configured based on the device's locale and region
+  /// settings.
+  ///
+  /// This method is private and used internally to manage platform patterns.
+  static void _setNumberSymbols(
+    String locale,
+    String icuNumberFormat,
+    String decimalSeparator,
+    String groupSeparator,
+  ) {
+    NumberSymbols localeSymbols =
+        numberFormatSymbols[locale] ?? numberFormatSymbols['en_US']!;
+    NumberSymbols regionSymbols = NumberSymbols(
+      NAME: _numberSymbolsName,
+      DECIMAL_SEP: decimalSeparator, // override with regional setting
+      GROUP_SEP: groupSeparator, // override with regional setting
+      PERCENT: localeSymbols.PERCENT,
+      ZERO_DIGIT: localeSymbols.ZERO_DIGIT,
+      PLUS_SIGN: localeSymbols.PLUS_SIGN,
+      MINUS_SIGN: localeSymbols.MINUS_SIGN,
+      EXP_SYMBOL: localeSymbols.EXP_SYMBOL,
+      PERMILL: localeSymbols.PERMILL,
+      INFINITY: localeSymbols.INFINITY,
+      NAN: localeSymbols.NAN,
+      DECIMAL_PATTERN: icuNumberFormat, // override with regional setting
+      SCIENTIFIC_PATTERN: localeSymbols.SCIENTIFIC_PATTERN,
+      PERCENT_PATTERN: localeSymbols.PERCENT_PATTERN,
+      CURRENCY_PATTERN: localeSymbols.CURRENCY_PATTERN,
+      DEF_CURRENCY_CODE: localeSymbols.DEF_CURRENCY_CODE,
+    );
+    numberFormatSymbols[_numberSymbolsName] = regionSymbols;
+  }
+
   /// Output regional settings as a formatted string.
   @override
   String toString() {
     return '''RegionSettings(
+      locale: $locale,
       temperatureUnits: $temperatureUnits,
       usesMetricSystem: $usesMetricSystem,
       firstDayOfWeek: $firstDayOfWeek,
       dateFormat: $dateFormat,
       numberFormat: $numberFormat,
+      icuNumberFormat: $icuNumberFormat,
+      decimalSeparator: $decimalSeparator,
+      groupSeparator: $groupSeparator,
     )''';
   }
 }